Fine tune build script to handle releases where release tags for gnucash and gnucash-docs differ
For example there can be a gnucash 3.8b tag with a gnucash-docs 3.8 tag.
The way to invoke the script for this is
build-package.sh -r3.8 -c3.8b -d3.8
or even
build-package.sh -r3.8 -c3.8b

The value of -r serves as both
- the default for -c and -d
- the sourceforge directory name containing the tarballs.

So the main constraint for release tarballs now is that the tarballs must live
in the same sourceforge directory.

Notes:
* if the release tag is the same for gnucash, gnucash-docs and the directory
(the ideal case), the command can be simplified even more:
build-package.sh -r3.7
* for development (nightly) builds the directory restriction doesn't apply. For
such builds you are free to use -c, -d and -r as you see fit. -r only serves
as default for -c and -d in that scenario.