Complete the manifest file modularization
- instead of combining templates into the final manifest file use a common
  manifest file for all common modules
- the modules that differ between tar or git builds will go into
  a dynamically generated module file gnucash.json
- the common manifest file always includes this file

Further, for easy lookup of which manifest and modules files were used
to generate a given flatpak, also upload the used module files to
the public download site. For example this can be used to quickly check
what version of aqbanking was built into a given flatpak.
As there are now more files to uplaod, store them in a subdirectory
manifests/{maint,master,...}/<branch>/{manifestfile,modules}