Relax default sandbox restrictions to allow filesystem access to the complete host system
This should make it easier for users to store there gnucash (meta)data outside
of their home directories, like on an external drive or a usb stick.
It may also help with using 'localhost' for mysql backed books but I haven't tested that.

If users want a tighter sandbox, they can override this default with the 'flatpak override' command