Remove dconf access
This is replaced with a settings (host to sandbox) migration option
which will only work if host has settings under the /org/gnucash/GnuCash tree
in GSettings. This will be as of GnuCash 4.7, but only if that version is run
natively on the host at least once. Until then, preferences migrations from
host to sandbox need a manual workaround as can be found here:
https://wiki.gnucash.org/wiki/Flatpak#Known_Issues