Fix quote-retrieval failure
The root problem was that Net::SSLeay had been removed from
the build because the script attempts to install all of the
modules. Net::SSLeay requires the libssl headers, requiring
in turn that on Linux systems the libssl-dev package is
installed. It was not installed on my system. Now it is and
Net::SSLeay is again included in the flatpak.

Fixes https://github.com/flathub/org.gnucash.GnuCash/issues/97